Create pageScript.js
My puny probably completely idiotic attempt at trying javascript. I've no real knowledge of it but I'd like to learn. This is my attempt at creating a script to update the links in a div when you press a button. Let's see if I can make this work... somehow.

This is just the absolute bare shell, I'll continually update this as I figure stuff out. (It's not even remotely functional or complete).